A couple snippets for markdown URLs/images.
Not terribly valuable; markdown is already simple. But at times people
forget the syntax for links so this helps.

There are other less common link markup types that seem to not fit as
snippets, so ignoring those.

I don’t know of a way to cancel out of “optional parameters”. E.g., the
title in markdown is optional, so if you don’t want one it would be nice
to back up to remove the quotes. But this is minor.